fix sign compare warnings
Do explicit casts before comparing signed with unsigned.  The one in
log.c actually fixes another warning: in the expanded macro, there may
be a check "logLevel>=0", which is always true.

git-svn-id: https://svn.musicpd.org/mpd/trunk@7230 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/interface.c
@@ -644,7 +644,7 @@ static void flushInterfaceBuffer(Interface * interface)
 		if (ret < 0)
 			break;
 		else if ((size_t)ret < buf->size) {
-			assert(interface->deferred_bytes >= ret);
+			assert(interface->deferred_bytes >= (size_t)ret);
 			interface->deferred_bytes -= ret;
 			buf->data = (char *)buf->data + ret;
 			buf->size -= ret;
